fix creating LIPS classes from SXML
The problem was code that was unboxing the values returned from Callback functions.
It was broken for Preact that called h function with lambda and then inside created
an instance of LIPS class. The problem was that constructor was wrapped with function
so the original created class (using prototypes) was lost and. The fix is to just
add the prototype from original function to wrapper